requirements.txt
streamlit
streamlit-clipboard
Pillow
chrono24
requests